+++ rally/common/utils.py
@@ -34,7 +34,6 @@ from six import moves
 
 from rally.common import logging
 from rally import exceptions
-from rally.utils import strutils
 
 LOG = logging.getLogger(__name__)
 
@@ -190,27 +189,6 @@ def first_index(lst, predicate):
         if predicate(e):
             return i
     return None
-
-
-@logging.log_deprecated(message="Its not used elsewhere in Rally already.",
-                        rally_version="0.4.1")
-def distance(s1, s2):
-    """Computes the edit distance between two strings.
-
-    The edit distance is the Levenshtein distance. The larger the return value,
-    the more edits are required to transform one string into the other.
-
-    :param s1: First string to compare
-    :param s2: Second string to compare
-    :returns: Integer distance between two strings
-    """
-    n = range(0, len(s1) + 1)
-    for y in range(1, len(s2) + 1):
-        l, n = n, [y]
-        for x in moves.range(1, len(s1) + 1):
-            n.append(min(l[x] + 1, n[-1] + 1,
-                         l[x - 1] + (s2[y - 1] != s1[x - 1])))
-    return n[-1]
 
 
 def retry(times, func, *args, **kwargs):
@@ -684,13 +662,6 @@ class LockedDict(dict):
     def clear(self, *args, **kwargs):
         self._check_is_unlocked()
         return super(LockedDict, self).clear(*args, **kwargs)
-
-
-@logging.log_deprecated(message="Its not used elsewhere in Rally already.",
-                        rally_version="0.11.2")
-def format_float_to_str(num):
-    """DEPRECATED. Use rally.utils.strutils.format_float_to_str instead."""
-    return strutils.format_float_to_str(num)
 
 
 class DequeAsQueue(object):