+++ rally/task/atomic.py
@@ -94,43 +94,6 @@ def action_timer(name):
     return wrap
 
 
-def optional_action_timer(name, argument_name="atomic_action", default=True):
-    """Optionally provide measure of execution time.
-
-    Decorates methods of the Scenario class. This provides duration in
-    seconds of each atomic action. When the decorated function is
-    called, this inspects the kwarg named by ``argument_name`` and
-    optionally sets an ActionTimer around the function call.
-
-    The ``atomic_action`` keyword argument does not need to be added
-    to the function; it will be popped from the kwargs dict by the
-    wrapper.
-
-    :param name: The name of the timer
-    :param argument_name: The name of the kwarg to inspect to
-                          determine if a timer should be set.
-    :param default: Whether or not to set a timer if ``argument_name``
-                    is not present.
-    """
-    def wrap(func):
-        @functools.wraps(func)
-        def func_atomic_actions(self, *args, **kwargs):
-            LOG.warning("'optional_action_timer' is deprecated "
-                        "since rally v0.10.0."
-                        "Please use action_timer instead, "
-                        "we have improved atomic actions, "
-                        "now do not need to explicitly close "
-                        "original action.")
-            if kwargs.pop(argument_name, default):
-                with ActionTimer(self, name):
-                    f = func(self, *args, **kwargs)
-            else:
-                f = func(self, *args, **kwargs)
-            return f
-        return func_atomic_actions
-    return wrap
-
-
 def merge_atomic_actions(atomic_actions, root=None, depth=0,
                          depth_of_processing=2):
     """Merge duplicates of atomic actions into one atomic action.