+++ tests/unit/common/test_utils.py
@@ -150,29 +150,6 @@ class FirstIndexTestCase(test.TestCase):
         self.assertIsNone(utils.first_index(lst, lambda e: e == 2))
 
 
-class EditDistanceTestCase(test.TestCase):
-
-    def test_distance_empty_strings(self):
-        dist = utils.distance("", "")
-        self.assertEqual(0, dist)
-
-    def test_distance_equal_strings(self):
-        dist = utils.distance("abcde", "abcde")
-        self.assertEqual(0, dist)
-
-    def test_distance_replacement(self):
-        dist = utils.distance("abcde", "__cde")
-        self.assertEqual(2, dist)
-
-    def test_distance_insertion(self):
-        dist = utils.distance("abcde", "ab__cde")
-        self.assertEqual(2, dist)
-
-    def test_distance_deletion(self):
-        dist = utils.distance("abcde", "abc")
-        self.assertEqual(2, dist)
-
-
 class TenantIteratorTestCase(test.TestCase):
 
     def test_iterate_per_tenant(self):